% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{myoscore_dimensions}
\alias{myoscore_dimensions}
\title{Get MyoScore dimension names}
\usage{
myoscore_dimensions()
}
\value{
Character vector of the five dimension names.
}
\description{
Get MyoScore dimension names
}
\examples{
myoscore_dimensions()
}
