% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{myoscore_weights}
\alias{myoscore_weights}
\title{Get MyoScore dimension weights}
\usage{
myoscore_weights()
}
\value{
Named numeric vector of dimension weights (sum to 1.0).
}
\description{
Get MyoScore dimension weights
}
\examples{
myoscore_weights()
}
